New: Added textSelection Notifications to CPTextField
This deviates from cocoa in that only NSTextViews send selection notification
and NSTextFields use a hidden NSTextView for its field editor... But we don't have CPTextViews.

The real motivation behind receiving notifications upon selection change is to enable
potential toolbar items that rely on textfield selection to be enabled/disabled.