Fix: don't show browser context menu in menubars
In cocoa, clicking the right mouse on the menubar activates the menu.
Our menubar is a subclass of a view which attempts to display a context submenu
on right click. And failing to find one passes on the click to the browser which
then displays its context menu. In Safari and Chrome it leaves all of the menu
items Highlighted makeing the App appear broken.

This fix simply forces rightclicks to be passed on as regular clicks in menubars.